Changed what de_end represents
This may be a mistake but it was trigger by trying to get real mode
working.  The hardware page tables start at ffffffffe0000000 and have a
length of 0000000020000000.  This makes start+length equal to 0.  de_len
use to be start+length but it was changed.  Now it is start+length-1 so
that the above case, de_end comes out to be ffffffffffffffff which sorts
properly.