Better real mode
I had the dump_entry with a union for the virtual entries and the real
entries.  The reason is a virtual entry is "long" while a real entry
is "long long" but they both come out to be the same size in a 64 bit
system.  The virtual entry had an de_end and a de_min.  The de_min was
key in the search of the dump entries but the real union didn't have
the same concept.

I decided to just smash the union together and have only one type of
entry but still keeping the de_isreal bit.

The root issue was the dump had disjoint page entries followed by a
big entry that encompassed the disjoint entries plus the pages between
them.  The original code looked at starting_d only but did not walk up
the list to see if there was an entry that would work.  To do the walk
up the list, I needed de_min so I would know when to stop.