Added pipe and file redirect
The changes are not complete but the basics of piping the output of a
statement or redirecting the output of a statement to a file was
added.  The scanner is perhaps an odd place to do this but anywhere
else and multiple tokens are needed before the statement resolves.
This makes interactive use a pain.

Part of the issue was fgets and readline trim off the newline from the
input.  This confused me because the scanner patterns were never
matching.

I also cleaned up adding strings to the string buffer and made it more
robust.