Fixed non_function_declarator
The result of "char *foo[15];" was a pointer to an array which is not
correct.  It should be an array of pointers.  This was caused when I
was prepending each new piece instead of appending it.

Also removed a debug mistake to load just the first of the TP_ARRAY as
well as added some debugging to help track down the above issue.